change structure nav-bar components
