+++ src/components/core/nav-bar-routes.vue
@@ -23,7 +23,7 @@ export default {
     // to return an array of children, rather than a single
     // root node.
     return props.routes.map(route => (
-      <VBtn flat key={route.name} to={route}>
+      <VBtn class="v-btn--simple" round color="default" key={route.name} to={route}>
         {getRouteTitle(route)}
       </VBtn>
     ))
