+++ src/components/core/nav-bar.vue
@@ -1,0 +1,48 @@
+<script>
+import NavBarRoutes from './nav-bar-routes'
+import Toolbar from './tool-bar'
+import Drawer from './drawer'
+
+export default {
+  components: { NavBarRoutes, Toolbar, Drawer },
+  data() {
+    return {
+      drawer: true,
+    }
+  },
+  computed: {
+    lock() {
+      // check ว่า เป็นหน้า home หรือ login รึเปล่า ถ้าไม่ใช่ return true
+      if (this.$route.name === 'home')
+        return false
+      if (this.$route.name === 'login')
+        return false
+
+      return true
+    }
+  },
+  methods: {
+    onResponsiveInverted () {
+      if (window.innerWidth < 991) {
+        this.responsive = true
+      } else {
+        this.responsive = false
+      }
+    }
+  }
+}
+</script>
+
+<template>
+  <v-layout>    
+    <toolbar />
+
+    <drawer v-if="lock"/>
+  </v-layout>
+</template>
+
+<style lang="scss" scoped>
+@import '@design';
+
+
+</style>
