+++ src/design/main.styl
@@ -1,4 +1,5 @@
 @import '../../node_modules/vuetify/src/stylus/settings/_colors'
+@import '../../node_modules/vuetify/src/stylus/main'
 
 $theme := {
   primary: '#33AC41'
@@ -10,4 +11,4 @@ $theme := {
   success: '#1C8B2D'
 }
 
-@import '../../node_modules/vuetify/src/stylus/main'
+
