+++ src/design/material-dashboard/_checkboxes.scss
@@ -1,3 +1,3 @@
-.mdi-checkbox-blank-outline:before {
+.mdi-checkbox-blank-outline::before {
   -webkit-text-stroke: 1px $white;
 }
