+++ src/design/material-dashboard/_colors.scss
@@ -45,7 +45,7 @@ $purple-900: #4a148c !default;
 $purple-A100: #ea80fc !default;
 $purple-A200: #e040fb !default;
 $purple-A400: #d500f9 !default;
-$purple-A700: #aa00ff !default;
+$purple-A700: #a0f !default;
 $purple: $purple-500 !default;
 
 // use this in SECONDARY COLOR
@@ -226,7 +226,7 @@ $yellow-700: #fbc02d !default;
 $yellow-800: #f9a825 !default;
 $yellow-900: #f57f17 !default;
 $yellow-A100: #ffff8d !default;
-$yellow-A200: #ffff00 !default;
+$yellow-A200: #ff0 !default;
 $yellow-A400: #ffea00 !default;
 $yellow-A700: #ffd600 !default;
 $yellow: $yellow-700 !default;
@@ -302,7 +302,7 @@ $brown: $brown-500 !default;
 
 $grey-50: #fafafa !default;
 $grey-100: #f5f5f5 !default;
-$grey-200: #eeeeee !default;
+$grey-200: #eee !default;
 $grey-300: #e0e0e0 !default;
 $grey-400: #bdbdbd !default;
 $grey-500: #9e9e9e;  $rgb-grey-500: "158, 158, 158" !default;
@@ -311,7 +311,7 @@ $grey-700: #616161 !default;
 $grey-800: #424242 !default;
 $grey-900: #212121 !default;
 $grey-A100: #f5f5f5 !default;
-$grey-A200: #eeeeee !default;
+$grey-A200: #eee !default;
 $grey-A400: #bdbdbd !default;
 $grey-A700: #616161 !default;
 $grey: $grey-500 !default;
@@ -334,5 +334,5 @@ $blue-grey-A700: #455a64 !default;
 $blue-grey: $blue-grey-500 !default;
 
 
-$black: #000000; $rgb-black: "0,0,0" !default;
-$white: #ffffff; $rgb-white: "255,255,255" !default;
+$black: #000; $rgb-black: "0,0,0" !default;
+$white: #fff; $rgb-white: "255,255,255" !default;
