+++ src/design/material-dashboard/_inputs.scss
@@ -1,9 +1,9 @@
-.theme--light.v-text-field .v-input__slot:before {
+.theme--light.v-text-field .v-input__slot::before {
   color: $input-color;
   border-color: $input-color !important;
 }
 
-.input--is-focused .v-input__control .v-input__slot:after{
+.input--is-focused .v-input__control .v-input__slot::after{
   border-color: red !important;
 }
 
@@ -15,16 +15,16 @@
   }
 
   .v-label {
-    color: $label-color !important;
     font-size: $font-size-default;
     font-weight: $font-weight-base;
+    color: $label-color !important;
   }
 }
 
 .theme--light.v-input:not(.v-input--is-disabled) input,
 .theme--light.v-input:not(.v-input--is-disabled) textarea {
-  color: $black-color;
   font-size: $font-size-default;
+  color: $black-color;
 }
 
 // ใช้กำหนดสี โดย กำหนดจาก class="purple-input"
