+++ src/design/material-dashboard/_mixins.scss
@@ -1,6 +1,6 @@
 @mixin shadow-big(){
   box-shadow: 0 10px 30px -12px rgba(0, 0, 0, $mdb-shadow-key-penumbra-opacity * 3),
-              0  4px 25px 0px rgba(0, 0, 0, $mdb-shadow-ambient-shadow-opacity),
+              0  4px 25px 0 rgba(0, 0, 0, $mdb-shadow-ambient-shadow-opacity),
               0  8px 10px -5px rgba(0, 0, 0, $mdb-shadow-key-umbra-opacity);
 }
 
@@ -12,7 +12,7 @@
 
 @mixin button-shadow-color($color){
     box-shadow: 0 14px 26px -12px rgba($color, $mdb-shadow-key-penumbra-opacity * 3),
-    0  4px 23px 0px rgba(0,0,0, $mdb-shadow-ambient-shadow-opacity),
+    0  4px 23px 0 rgba(0,0,0, $mdb-shadow-ambient-shadow-opacity),
     0  8px 10px -5px rgba($color, $mdb-shadow-key-umbra-opacity) !important;
 }
 
@@ -24,7 +24,7 @@
 
 @mixin shadow-big-color($color){
   box-shadow: 0 12px 20px -10px rgba($color, $mdb-shadow-key-penumbra-opacity * 2),
-              0  4px 20px 0px rgba(0,0,0, $mdb-shadow-ambient-shadow-opacity),
+              0  4px 20px 0 rgba(0,0,0, $mdb-shadow-ambient-shadow-opacity),
               0  7px 8px -5px rgba($color, $mdb-shadow-key-umbra-opacity) !important;
 }
 
@@ -35,7 +35,7 @@
 }
 
 @mixin inputs-color($color, $color-name) {
-  .v-text-field.#{$color-name}-input > .v-input__control > .v-input__slot:after {
+  .v-text-field.#{$color-name}-input > .v-input__control > .v-input__slot::after {
     border-color: $color;
   }
 }
@@ -43,6 +43,7 @@
 @mixin drawer-active-colors($color-name, $color) {
   .v-navigation-drawer .v-list .v-list-item .v-list__tile.v-list__tile--active.#{$color-name} {
     background-color: $color !important;
+
     @include shadow-big-color($color);
   }
 }
@@ -50,6 +51,7 @@
 @mixin toolbar-colors($color-name, $color) {
   &.#{$color-name} {
     background-color: $color !important;
+
     @include shadow-big-color($color);
   }
 }
@@ -70,6 +72,7 @@
 @mixin card-header-color($color, $shadow-color, $color-400, $color-600) {
   .v-card.#{$color} {
     background: linear-gradient(60deg, $color-400, $color-600) !important;
+
     @include shadow-big-color($shadow-color);
   }
 }
@@ -86,11 +89,13 @@
 
 @mixin dropdown-menu-color($color) {
   background-color: $color !important;
+
   @include button-shadow-color($color);
 }
 
 @mixin button-color($color) {
   background-color: $color !important;
+
   @include shadow-2dp-color($color);
 
   &:focus,
@@ -98,6 +103,7 @@
   &:hover{
     // remove this line if you want black shadows
     @include button-shadow-color($color);
+
     background-color: darken($color, 3%) !important;
   }
 
@@ -111,8 +117,8 @@
       &:active:hover,
       &.active:focus,
       &.active:hover{
-        background-color: transparent !important;
         color: $color !important;
+        background-color: transparent !important;
         box-shadow: none !important;
 
         i{
@@ -123,8 +129,9 @@
 }
 
 @mixin social-buttons-color ($color){
-    background-color: $color !important;
     color: $white;
+    background-color: $color !important;
+
     @include shadow-2dp-color($color);
 
     &:hover,
@@ -135,8 +142,9 @@
     &:active:hover,
     &.active:focus,
     &.active:hover{
-        background-color: $color !important;
         color: $white;
+        background-color: $color !important;
+
         @include button-shadow-color($color);
     }
