+++ src/design/material-dashboard/_tables.scss
@@ -1,6 +1,6 @@
 .theme--light.v-table tbody tr:not(:last-child),
 .theme--light.v-table thead tr:first-child {
-  border-bottom: 1px solid $table-border-color!important;
+  border-bottom: 1px solid $table-border-color !important;
 }
 
 table.v-table thead tr {
