+++ src/design/material-dashboard/_tabs.scss
@@ -1,7 +1,7 @@
 .card-tabs {
   .v-divider {
-    margin-left: 0 !important;
     margin-right: 0 !important;
+    margin-left: 0 !important;
   }
 
   .flex {
@@ -22,12 +22,12 @@
   }
 
   .v-list__tile__title {
-    font-size: $font-size-default;
-    font-weight: $font-weight-lighter;
+    height: auto;
     padding: 0 $tabs-text-box-padding;
     overflow: visible;
+    font-size: $font-size-default;
+    font-weight: $font-weight-lighter;
     white-space: inherit;
-    height: auto;
   }
 
   .v-list .v-input, .v-list .v-input__slot {
